xeet: remove internal output comparison
Remove all the functionality for output comparison form Xeet's
infrastructure. This is deferred to the tests writer, as output
comparison is only suitable for very specific cases, and quite needs to
be further extended anyway. On the other hand, it removes a lot of code
from the Xeet code base, making it easier to handle

The result a rather large change in the way xeet tests itself. Instead
of relying on its own output compare mechanism, a specific xeet
stdout/stderr comparison is used.

Not that debug mode prevents output files from being created at the
moment, so verification based on these files is not supported when using
debug mode. The xeet self testing just skip the verification.

This isn't a degradation, as this was the previous behavior, but this
change streamline and makes this it easier to understand.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>