tests: rewrite
Put some order in the testing flow of the module. In general, xeet uses
a previous version of xeet (one that is probably checked, hopefully),
to test itself. That's the gist of it. A few files are in play here:

1. runtests.sh - runs the suite of tests using a xeet binary out of the
   module virtual environment, probably the system installed one. It
   calls xeet with the xeet.json located at the tests directory.

2. runtests_internal.sh - script used the xeet.json file used by
   runtests. Unless stated otherwise, it activates the virtual
   environment in the repository root, and run a test. Each test is
   actually a xeet invocation, but the one in development.

   If a special environment variable named XEET_PKG_VENV_PATH is set,
   runtests_internal.sh will not use the repository virtual environment,
   but the one set in the variable. This is used mostly for packaging
   tests.

3. run_pkg_tests.sh - for testing packages. Create a separated virtual
   environment, installs xeet from pypi (or it's test counterpart), sets
   XEET_PKG_VENV_PATH to that environment path, and calls runtests.sh.
   So the external xeet is still the system xeet being ran, but xeet in
   the tests themselves is the one in the newly created virtual
   environment.

The tests themselves are the numbered directory. Note that while the
external xeet.json tests should all pass, some of the internal tests
should fail, as tests failures are an important aspect of the test.
Hence - the outer xeet validates that the internal xeet behaves as
expected, including failures.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>