xeet: xstep
Part 2 of the rehaul patch set. Separate the tests and the actual steps
it takes:

Each test is constructed by up to 3 lists of steps. A step is a unit
that can be executed and can have 2 dimensions to its execution:
1. Completed - if the step had successfully finished
2. Failed - if the step had failed

These 2 are distinct. An example can be with the 1st step type which is
the execute binary step: if the command fails to even start (e.g. No
such file or so), the step is not completed. If the step does complete
but doesn't meet the required conditions for success (valid exit code,
for example), it is considered a failure.

A test can have 3 types of the steps list:
1. Run - If any of the steps fails to complete, the test is considered
   to have run error. If all steps completed and any of them failed, the
   test is considered a failure. Otherwise, the test has passed. The run
   list isn't ran if the pre-run list has failed to run. This list
   should be used for setting the test environment if such is needed.
2. Pre-run - Steps to be ran before the run list. If any of the steps in
   it fail or do not complete, the test status is set to run error, and
   the run list isn't invoked.
3. Post-test - Steps that are ran after the pre-run and run list are
   finished. This list is always ran, and doesn't affect the test
   status. Should be used for cleanup.

This patch also introduces 2 initial steps to begin with:
1. Dummy step for testing purposes.
2. Exec step to run binaries, practically moving the binary execution
   logic from the previous test into the step.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>