core: resource synchronization
Add the implementation of the resource management and synchronization
between tests. A test might require a resource and will not run without
its acquisition. If a test require a resource that isn't available at the
moment due to other thread running a test with the resource acquired,
the next test is attempted, and the busy resource test will be attempted
later, checking the resource requirement again.

At some point, the acquiring test will finish and the resource will be
freed for other tests usage.

There are many subtleties regarding the thread synchronization over the
resource not detailed here.

Resource might or might not represent a "real" resource. Threads are
always synchronized over the resources they require, so a "virtual"
resource might be used for synchronizing tests that shouldn't run in
parallel for any reason.

When a resource is acquired, it becomes a Xvar available for test setup
like any other var, and since the resource (like other variables) can be
and json-like structure, the resource details are available for the test.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>