xeet: global matrix support
Add matrix variable at global level. A matrix variable is a list-based
variable deriving an iteration with each of the possible values in the
list. Multiple matrix variables create an iteration with each possible
matrix permutation.

The change is rather big as now each iteration run all the possible
matrix permutations in it (1 if no matrix is defined), so it affects
existing unit tests.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>